requirements.txt
opencv-python>=4.8.1
numpy>=1.26.3
mysql-connector-python>=8.3.0
python-dotenv>=1.0.1
redis>=5.0.1
scikit-learn>=1.4.0
fastapi>=0.109.2
uvicorn>=0.27.1
requests>=2.31.0
python-multipart>=0.0.7
tenacity>=8.2.3